src/requirements.txt
Flask==2.1.2
pyyaml==6.0
